oc: re-use edit from kubectl
