+++ contrib/completions/bash/oc
@@ -1650,6 +1650,7 @@ _oc_edit()
     flags+=("--output=")
     two_word_flags+=("-o")
     flags+=("--output-version=")
+    flags+=("--save-config")
     flags+=("--windows-line-endings")
     flags+=("--alsologtostderr")
     flags+=("--api-version=")
