+++ contrib/completions/bash/openshift
@@ -6715,6 +6715,7 @@ _openshift_cli_edit()
     flags+=("--output=")
     two_word_flags+=("-o")
     flags+=("--output-version=")
+    flags+=("--save-config")
     flags+=("--windows-line-endings")
     flags+=("--alsologtostderr")
     flags+=("--api-version=")
