+++ pkg/cmd/cli/cmd/wrappers.go
@@ -509,3 +509,45 @@ func NewCmdConvert(fullName string, f *clientcmd.Factory, out io.Writer) *cobra.
 	cmd.Example = fmt.Sprintf(convertExample, fullName)
 	return cmd
 }
+
+const (
+	editLong = `
+Edit a resource from the default editor
+
+The edit command allows you to directly edit any API resource you can retrieve via the
+command line tools. It will open the editor defined by your OC_EDITOR, or EDITOR environment
+variables, or fall back to 'vi' for Linux or 'notepad' for Windows. You can edit multiple
+objects, although changes are applied one at a time. The command accepts filenames as well
+as command line arguments, although the files you point to must be previously saved versions
+of resources.
+
+The files to edit will be output in the default API version, or a version specified
+by --output-version. The default format is YAML - if you would like to edit in JSON
+pass -o json. The flag --windows-line-endings can be used to force Windows line endings,
+otherwise the default for your operating system will be used.
+
+In the event an error occurs while updating, a temporary file will be created on disk
+that contains your unapplied changes. The most common error when updating a resource
+is another editor changing the resource on the server. When this occurs, you will have
+to apply your changes to the newer version of the resource, or update your temporary
+saved copy to include the latest resource version.`
+
+	editExample = `  # Edit the service named 'docker-registry':
+  $ %[1]s edit svc/docker-registry
+
+  # Edit the DeploymentConfig named 'my-deployment':
+  $ %[1]s edit dc/my-deployment
+
+  # Use an alternative editor
+  $ OC_EDITOR="nano" %[1]s edit dc/my-deployment
+
+  # Edit the service 'docker-registry' in JSON using the v1beta3 API format:
+  $ %[1]s edit svc/docker-registry --output-version=v1beta3 -o json`
+)
+
+func NewCmdEdit(fullName string, f *clientcmd.Factory, out io.Writer) *cobra.Command {
+	cmd := kcmd.NewCmdEdit(f.Factory, out)
+	cmd.Long = editLong
+	cmd.Example = fmt.Sprintf(editExample, fullName)
+	return cmd
+}
