--- pkg/cmd/util/editor/term_unsupported.go
@@ -1,10 +1,0 @@
-// +build windows
-
-package editor
-
-import (
-	"os"
-)
-
-// childSignals are the allowed signals that can be sent to children in Windows to terminate
-var childSignals = []os.Signal{os.Interrupt}
